pyproject.toml
[tool.poetry]
name = "test-project"
version = "0.1.0"
description = ""
authors = ["Nursultan Kudaibergenov <aka.tworedz@gmail.com>"]
readme = "README.md"
packages = [{include = "test_project"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.7"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.25.1"
django-4-jet = "^1.0.9"
django-silk = "^5.0.3"
pillow = "^9.4.0"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
pytest = "^7.2.1"
pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
